+++ src/main/java/com/epam/reportportal/utils/ParameterUtils.java
@@ -17,7 +17,7 @@
 package com.epam.reportportal.utils;
 
 import com.epam.reportportal.annotations.ParameterKey;
-import com.epam.reportportal.utils.markdown.MarkdownUtils;
+import com.epam.reportportal.utils.formatting.MarkdownUtils;
 import com.epam.ta.reportportal.ws.model.ParameterResource;
 import org.apache.commons.lang3.tuple.Pair;